+++ config/urls.py
@@ -1,7 +1,8 @@
 from django.conf import settings
 from django.conf.urls import include
 from django.contrib import admin
-from django.urls import path
+from django.urls import path, reverse_lazy
+from django.views.generic import RedirectView
 
 from wagtail.admin import urls as wagtailadmin_urls
 from wagtail.core import urls as wagtail_urls
@@ -22,6 +23,13 @@ urlpatterns = [
     # the list:
     path('', include(wagtail_urls)),
 ]
+
+
+if settings.ENFORCE_STAFF_SSO_ENABLED:
+    urlpatterns += [
+        path('admin/login/', RedirectView.as_view(url=reverse_lazy('authbroker_client:login'), query_string=True)),
+        path('auth/', include('authbroker_client.urls')),
+    ]
 
 
 if settings.DEBUG: