+++ core/models.py
@@ -1,6 +1,8 @@
 import hashlib
 from django.db import models
 
+from wagtail.images.models import Image, AbstractImage, AbstractRendition
+
 
 class AbstractObjectHash(models.Model):
     class Meta:
@@ -35,3 +37,16 @@ class ImageHash(AbstractObjectHash):
         on_delete=models.CASCADE,
         related_name='+'
     )
+
+
+class AltTextImage(AbstractImage):
+    alt_text = models.CharField(max_length=255, blank=True)
+
+    admin_form_fields = Image.admin_form_fields + ('alt_text',)
+
+
+class Rendition(AbstractRendition):
+    image = models.ForeignKey(AltTextImage, on_delete=models.CASCADE, related_name='renditions')
+
+    class Meta:
+        unique_together = (('image', 'filter_spec', 'focal_point_key'))