+++ domestic/blocks.py
@@ -1,10 +1,13 @@
 from wagtail.core import blocks
+from wagtail.images.blocks import ImageChooserBlock
+
 from core import blocks as core_blocks
 
 
-class CampaignBlock(core_blocks.ImageBaseBlock, core_blocks.LinkBlock):
+class CampaignBlock(core_blocks.LinkBlock):
     heading = blocks.CharBlock()
     subheading = blocks.CharBlock()
+    image = ImageChooserBlock()
 
 
 class MarketAccessDBBlock(core_blocks.LinkWithImageAndContentBlock):