+++ CHANGELOG.md
@@ -5,4 +5,4 @@
 ### Fixed bugs
 
 ### Implemented enhancements
-- MVP-79 - Add react built tool
+- MVP-79 - Add react build tool